requirements.txt
Flask
firebase-admin
requests
python-dotenv
scikit-learn
numpy